Add local thread store archive migration
