+++ codex-rs/rollout/src/lib.rs
@@ -48,6 +48,7 @@ pub use list::get_threads_in_root;
 pub use list::parse_cursor;
 pub use list::read_head_for_summary;
 pub use list::read_session_meta_line;
+pub use list::read_thread_item_from_rollout;
 pub use list::rollout_date_parts;
 pub use metadata::builder_from_items;
 pub use policy::EventPersistenceMode;
